feat: Player, Dealer -> Participants로 추상화
