+++ src/main/java/domain/model/Player.java
@@ -1,35 +1,8 @@
 package domain.model;
 
-import java.util.Set;
-
-public class Player {
-
-    private final Cards cards;
-    private Score score;
+public class Player extends Participant {
 
     public Player(final Cards cards) {
-        this.cards = cards;
-        this.score = makeScore(cards);
-    }
-
-    private Score makeScore(final Cards cards) {
-        return Score.of(cards);
-    }
-
-    public void addCard(final Card card) {
-        cards.add(card);
-        score = makeScore(cards);
-    }
-
-    public boolean isBust() {
-        return score.isBust();
-    }
-
-    public Cards getCards() {
-        return new Cards(Set.copyOf(cards.getCards()));
-    }
-
-    public Score getScore() {
-        return score;
+        super(cards);
     }
 }
